Extract MarkdownFences utility from SonicPiAgent
Move the fence-stripping logic into com.embabel.demo.util.MarkdownFences
so it can be reused by other agents that round-trip code or markdown
through the LLM. Move the corresponding unit test to
com.embabel.demo.util.MarkdownFencesTest.
